db_optimizer/requirements.txt
psycopg2-binary
numpy
pandas